refactored service methods and repository class for an int Person primary key

+++ fitness-tracker-backend/src/main/java/com/willcb/fitnesstrackerbackend/repositories/PersonRepository.java
@@ -6,7 +6,8 @@ import org.springframework.stereotype.Repository;
 import com.willcb.fitnesstrackerbackend.entities.Person;
 
 @Repository
-public interface PersonRepository extends JpaRepository<Person, Long> {
+public interface PersonRepository extends JpaRepository<Person, Integer> {
     // put queries here?
+    
 }
